requirements.txt
certifi==2018.4.16
future==0.16.0
python-dotenv==0.9.0
python-telegram-bot==10.1.0
